Introduce consumer-driven contract testing using Pact (#352)
* Implement first basic contract test using pact

* Fix lint errors

* Integrate CDCT into GitHub Actions Workflow
This approach uses the local file system for contract exchange

* Fix error in CDCT-related docker files
The contract files are now properly shared between providers and consumers.
Previously, the provider did not use the newly generated contract files.

* Delete all contract files
This commit serves as a test to see if GitHub Actions commits the auto-generated contract files

* Update auto-generated contracts

* Create separate GitHub Actions Workflow for CDCT

* Update auto-generated contracts

* Improve mocking in pipeline provider test
Now environment variables are properly mocked
Also the main method is invoked to start the provider
This removes the duplicate code in the provider test

* Change pactfile write mode to overwrite
Now existing pact files are entirely overwritten when consumer test execution brings changes to the files

Co-authored-by: felix-oq <felix-oq@users.noreply.github.com>