Fix non-production images to be uploaded as artifacts (#372)
* Implement first basic contract test using pact

* Fix lint errors

* Integrate CDCT into GitHub Actions Workflow
This approach uses the local file system for contract exchange

* Fix error in CDCT-related docker files
The contract files are now properly shared between providers and consumers.
Previously, the provider did not use the newly generated contract files.

* Delete all contract files
This commit serves as a test to see if GitHub Actions commits the auto-generated contract files

* Update auto-generated contracts

* Create separate GitHub Actions Workflow for CDCT

* Update auto-generated contracts

* Improve mocking in pipeline provider test
Now environment variables are properly mocked
Also the main method is invoked to start the provider
This removes the duplicate code in the provider test

* Change pactfile write mode to overwrite
Now existing pact files are entirely overwritten when consumer test execution brings changes to the files

* Add CDCT for all interactions of UI and Pipeline

* Replace static call to TransformationRest
It is necessary to call the constructor and pass the pipeline service url

* Add global test setup for UI
Now when running CDCT, the previously existing contract files are deleted.
Without this change, deprecated interactions would have to be deleted manually.

* Fix error in Dockerfile.consumer of UI
Previously the Dockerfile did not copy the pact folder.

* Update auto-generated contracts

* Move all interactions into a single test file
With the previous approach, the order of the interactions in the contract file could be random.
Now the order of the interactions in the contract file is determined by the order of the test cases.

* Update auto-generated contracts

* Update auto-generated contracts

* Delete deprecated pact test
This file was unintentionally added during merge resolution

* Update auto-generated contracts

* Remove obsolete common fixtures file

* Implement first message-based CDC-Test
The test covers an interaction between the notification service and the pipeline service

* Update auto-generated contracts

* Fix errors in GitHub Actions Workflow for CDCT
Now the Consumer-Side fails if any Consumer-Test fails.
The Provider-Side should run properly now.

* Update auto-generated contracts

* Make CDCT run in a single Workflow Job
Previously newly generated contract files were not available to the Provider-Side Job.

* Update auto-generated contracts

* Add jest config file for pipeline provider test

* Integrate CDCT into the regular ODS workflow
Pacts are no longer committed to the repository. They are uploaded as an artifact instead.
No separate CDCT workflow is required anymore.
Reduces the number of docker-compose files by introducing additional stages in dockerfiles and overriding docker-compose files for testing.

* Add build command to before running tests

* Fix building step of integration test in notification

* Fix command for provider-side test execution

* Add new environment variable to test setup

* Fix lint errors

* Add bash scripts for executing CDC-Tests

* Add Readme for CDCT

* Add comparison to integration tests in CDCT Readme

* Move all CDCT test files into their src directories
This makes the location of the test files uniform and easier to find

* Resolve merge conflict with dockerfiles

* Extend build stage in dockerfiles
Linting is now moved to the build stage to further speed up the building of the base stage.
Unit test execution is added to the build stage again. This ensures, that docker images can only be built, if their respective unit test is passing.

* Improve CDCT test scripts
The scripts abort now, if any command fails.
The script for the notification serivce now uses the scale option to prevent dependent but unnecessary services from being started during CDCT.

* Use script files for CDCT in workflow

* Fix exit codes of CDCT-scripts

* Use alternative option to don't start dependent services
Uses --no-deps instead of possibly multiple --scale options

* Mark scripts for CDCT executable

* Fix error in CDCT scripts
The CDCT scripts can now be run from any directory.

* Delete docker-compose.test.yml
It is currently unnecessary to run unit tests with Docker

* Execute CDCT after uploading docker images

* Fix duplicate CDCT step due to faulty merge

* Fix another duplicate CDCT step due to faulty merge

* Add an explanation why CDC-Tests are executed after image upload

* Remove unnecessary build for integration test of notification
Due to the step reordering, the notification image does not have to be rebuilt for the integration test

Co-authored-by: felix-oq <felix-oq@users.noreply.github.com>